+++ cbits/sha256.h
@@ -27,6 +27,7 @@
 #define CRYPTOHASH_SHA256_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 struct sha256_ctx
 {
@@ -39,7 +40,7 @@ struct sha256_ctx
 #define SHA256_CTX_SIZE		sizeof(struct sha256_ctx)
 
 void hs_cryptohash_sha256_init (struct sha256_ctx *ctx);
-void hs_cryptohash_sha256_update (struct sha256_ctx *ctx, const uint8_t *data, uint32_t len);
+void hs_cryptohash_sha256_update (struct sha256_ctx *ctx, const uint8_t *data, size_t len);
 void hs_cryptohash_sha256_finalize (struct sha256_ctx *ctx, uint8_t *out);
 
 #endif
